refactor: refactor processing in ManagerBase to ProcessorHelper class
